+++ tests/test_config_plugin.py
@@ -81,14 +81,14 @@ def test_plugin_instance():
             continue
         obj = plugin.get_instance(each['name'])
         for k,v in list(each.items()):
-            assert v == obj.config[k]
+            assert v == obj.pluginmgr_config[k]
 
     obj = plugin.get_instance(anon_config)
-    assert obj.config == anon_config
+    assert obj.pluginmgr_config == anon_config
 
     # copy ctor
     obj = plugin.get_instance('fancy_copy')
-    assert 'reeb' == obj.config['str0']
+    assert 'reeb' == obj.pluginmgr_config['str0']
 
     obj = plugin.get_instance({'fancy_copy': {'var0': 'luggage'}})
-    assert 'reeb' == obj.config['str0']
+    assert 'reeb' == obj.pluginmgr_config['str0']