+++ tests/test_varargs.py
@@ -25,7 +25,7 @@ kwargs = {'feast': 'lambs', 'and': 'cereals'}
 def test_no_varargs():
     Static0 = plugin.get_plugin_class('static0')
     obj = Static0(config)
-    assert config == obj.config
+    assert config == obj.pluginmgr_config
     assert not obj.args
     assert not obj.kwargs
 
@@ -33,7 +33,7 @@ def test_no_varargs():
 def test_args():
     Static0 = plugin.get_plugin_class('static0')
     obj = Static0(config, 1, 2, 5, 'three sir')
-    assert config == obj.config
+    assert config == obj.pluginmgr_config
     assert args == obj.args
     assert not obj.kwargs
 
@@ -41,7 +41,7 @@ def test_args():
 def test_kwargs():
     Static0 = plugin.get_plugin_class('static0')
     obj = Static0(config, **kwargs)
-    assert config == obj.config
+    assert config == obj.pluginmgr_config
     assert not obj.args
     assert kwargs == obj.kwargs
 
@@ -49,28 +49,28 @@ def test_kwargs():
 def test_both():
     Static0 = plugin.get_plugin_class('static0')
     obj = Static0(config, *args, **kwargs)
-    assert config == obj.config
+    assert config == obj.pluginmgr_config
     assert args == obj.args
     assert kwargs == obj.kwargs
 
 
 def test_new_plugin_type():
     obj = plugin.new_plugin(typconf, *args, **kwargs)
-    assert typconf == obj.config
+    assert typconf == obj.pluginmgr_config
     assert args == obj.args
     assert kwargs == obj.kwargs
 
 
 def test_new_plugin():
     obj = plugin.new_plugin({'static0': config}, *args, **kwargs)
-    assert config == obj.config
+    assert config == obj.pluginmgr_config
     assert args == obj.args
     assert kwargs == obj.kwargs
 
 
 def test_get_instance_anonymous():
     obj = plugin.get_instance(typconf, *args, **kwargs)
-    assert typconf == obj.config
+    assert typconf == obj.pluginmgr_config
     assert args == obj.args
     assert kwargs == obj.kwargs
 
@@ -80,7 +80,7 @@ def test_instantiate():
     nconf.update({'name': 'test_instantiate'})
     plugin.instantiate([{'static0': nconf}], *args, **kwargs)
     obj = plugin.get_instance('test_instantiate')
-    assert nconf == obj.config
+    assert nconf == obj.pluginmgr_config
     assert args == obj.args
     assert kwargs == obj.kwargs
 
@@ -88,7 +88,7 @@ def test_instantiate():
     ntype_conf = nconf.copy()
     ntype_conf.update({'type': 'test_instantiate'})
     obj = plugin.new_plugin(ntype_conf, *args, **kwargs)
-    assert ntype_conf == obj.config
+    assert ntype_conf == obj.pluginmgr_config
     assert args == obj.args
     assert kwargs == obj.kwargs