+++ CHANGELOG.md
@@ -8,9 +8,13 @@
 ### Fixed
 ### Changed
 - moved config plugins into separate config.ConfigPluginManager class
+- renamed plugin config attr to pluginmgr_config
 
 ### Deprecated
 ### Removed
+- name attr from ConfigPlugin objects
+- args and kwargs from config.PluginBase
+
 ### Security